+++ native/src/main/java/io/ballerina/stdlib/constraint/Constants.java
@@ -26,6 +26,14 @@ public class Constants {
     public static final String PREFIX_RECORD_FILED = "$field$";
     public static final String PREFIX_ANNOTATION_RECORD = "ballerina/constraint:1:";
 
+    public static final String SYMBOL_DOLLAR_SIGN = "$";
+    public static final String SYMBOL_DOT = ".";
+    public static final String SYMBOL_SINGLE_QUOTE = "'";
+    public static final String SYMBOL_COMMA = ",";
+    public static final String SYMBOL_OPEN_SQUARE_BRACKET = "[";
+    public static final String SYMBOL_CLOSE_SQUARE_BRACKET = "]";
+    public static final String SYMBOL_SEPARATOR = ":";
+
     public static final String ANNOTATION_TAG_INT = "Int";
     public static final String ANNOTATION_TAG_FLOAT = "Float";
     public static final String ANNOTATION_TAG_NUMBER = "Number";