+++ native/src/main/java/io/ballerina/stdlib/constraint/Constraints.java
@@ -30,9 +30,11 @@ import io.ballerina.stdlib.constraint.annotations.AbstractAnnotations;
 import io.ballerina.stdlib.constraint.annotations.RecordFieldAnnotations;
 import io.ballerina.stdlib.constraint.annotations.TypeAnnotations;
 
-import java.util.HashSet;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
-import java.util.Set;
+
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_DOLLAR_SIGN;
 
 /**
  * Extern functions for validating constraints.
@@ -40,12 +42,12 @@ import java.util.Set;
 public class Constraints {
 
     public static Object validate(Object value, BTypedesc typedesc) {
-        HashSet<String> failedConstraints = new HashSet<>();
+        List<String> failedConstraints = new ArrayList<>();
         try {
             Type type = typedesc.getDescribingType();
             if (type instanceof AnnotatableType) {
                 AbstractAnnotations annotations = getAnnotationImpl(type, failedConstraints);
-                annotations.validate(value, (AnnotatableType) type);
+                annotations.validate(value, (AnnotatableType) type, SYMBOL_DOLLAR_SIGN);
             } else if (type instanceof UnionType) {
                 Optional<Type> matchingType = getMatchingType(value, type);
                 if (matchingType.isPresent()) {
@@ -78,7 +80,7 @@ public class Constraints {
         return Optional.empty();
     }
 
-    private static AbstractAnnotations getAnnotationImpl(Type type, Set<String> failedConstraints) {
+    private static AbstractAnnotations getAnnotationImpl(Type type, List<String> failedConstraints) {
         if (type instanceof RecordType) {
             return new RecordFieldAnnotations(failedConstraints);
         } else {