+++ native/src/main/java/io/ballerina/stdlib/constraint/annotations/RecordFieldAnnotations.java
@@ -28,39 +28,43 @@ import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BDecimal;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.stdlib.constraint.Constants;
 
+import java.util.List;
 import java.util.Map;
-import java.util.Set;
+
+import static io.ballerina.stdlib.constraint.Constants.PREFIX_RECORD_FILED;
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_CLOSE_SQUARE_BRACKET;
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_DOT;
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_OPEN_SQUARE_BRACKET;
 
 /**
  * Extern functions for validating constraints on record fields.
  */
 public class RecordFieldAnnotations extends AbstractAnnotations {
 
-    private final Set<String> failedConstraints;
+    private final List<String> failedConstraints;
 
-    public RecordFieldAnnotations(Set<String> failedConstraints) {
+    public RecordFieldAnnotations(List<String> failedConstraints) {
         super(failedConstraints);
         this.failedConstraints = failedConstraints;
     }
 
     @SuppressWarnings("unchecked")
     @Override
-    public void validate(Object value, AnnotatableType type) {
+    public void validate(Object value, AnnotatableType type, String path) {
         BMap<BString, Object> record = (BMap<BString, Object>) value;
         BMap<BString, Object> recordAnnotations = type.getAnnotations();
         for (Map.Entry<BString, Object> entry : recordAnnotations.entrySet()) {
-            if (entry.getKey().getValue().startsWith(Constants.PREFIX_RECORD_FILED)) {
-                String fieldName = entry.getKey().getValue().substring(Constants.PREFIX_RECORD_FILED.length() + 1);
+            if (entry.getKey().getValue().startsWith(PREFIX_RECORD_FILED)) {
+                String fieldName = entry.getKey().getValue().substring(PREFIX_RECORD_FILED.length() + 1);
                 BMap<BString, Object> recordFieldAnnotations = (BMap<BString, Object>) entry.getValue();
                 Object fieldValue = getFieldValue(record, fieldName);
                 if (fieldValue != null) { // This can be null due to optional fields
-                    super.validateAnnotations(recordFieldAnnotations, fieldValue);
+                    super.validateAnnotations(recordFieldAnnotations, fieldValue, path + SYMBOL_DOT + fieldName);
                 }
             }
         }
-        validateReferredType(record, (RecordType) type);
+        validateReferredType(record, (RecordType) type, path);
     }
 
     private Object getFieldValue(BMap<BString, Object> record, String fieldName) {
@@ -75,7 +79,7 @@ public class RecordFieldAnnotations extends AbstractAnnotations {
         return obj;
     }
 
-    private void validateReferredType(BMap<BString, Object> record, RecordType type) {
+    private void validateReferredType(BMap<BString, Object> record, RecordType type, String path) {
         for (Field recordField : type.getFields().values()) {
             Type fieldType = recordField.getFieldType();
             String fieldName = recordField.getFieldName();
@@ -83,31 +87,34 @@ public class RecordFieldAnnotations extends AbstractAnnotations {
             if (fieldValue != null) { // This can be null due to optional fields
                 if (fieldType instanceof AnnotatableType) {
                     if (fieldType instanceof RecordType) {
-                        validate(fieldValue, (AnnotatableType) fieldType);
+                        validate(fieldValue, (AnnotatableType) fieldType, path + SYMBOL_DOT + fieldName);
                     } else {
                         TypeAnnotations typeAnnotations = new TypeAnnotations(this.failedConstraints);
-                        typeAnnotations.validate(fieldValue, (AnnotatableType) fieldType);
+                        typeAnnotations.validate(fieldValue, (AnnotatableType) fieldType,
+                                path + SYMBOL_DOT + fieldName);
                     }
                 } else if (fieldType instanceof ArrayType) {
-                    validateArrayType((ArrayType) fieldType, (BArray) fieldValue);
+                    validateArrayType((ArrayType) fieldType, (BArray) fieldValue, path + SYMBOL_DOT + fieldName);
                 }
             }
         }
     }
 
     @SuppressWarnings("unchecked")
-    private void validateArrayType(ArrayType fieldType, BArray fieldValue) {
+    private void validateArrayType(ArrayType fieldType, BArray fieldValue, String path) {
         Type elementType = fieldType.getElementType();
         if (elementType instanceof AnnotatableType) {
             if (elementType instanceof RecordType) {
                 for (int i = 0; i < fieldValue.getLength(); i++) {
                     BMap<BString, Object> map = (BMap<BString, Object>) fieldValue.getRefValue(i);
-                    validate(map, (AnnotatableType) elementType);
+                    validate(map, (AnnotatableType) elementType,
+                            path + SYMBOL_OPEN_SQUARE_BRACKET + i + SYMBOL_CLOSE_SQUARE_BRACKET);
                 }
             } else {
                 TypeAnnotations typeAnnotations = new TypeAnnotations(this.failedConstraints);
                 for (int i = 0; i < fieldValue.getLength(); i++) {
-                    typeAnnotations.validate(fieldValue.getRefValue(i), (AnnotatableType) elementType);
+                    typeAnnotations.validate(fieldValue.getRefValue(i), (AnnotatableType) elementType,
+                            path + SYMBOL_OPEN_SQUARE_BRACKET + i + SYMBOL_CLOSE_SQUARE_BRACKET);
                 }
             }
         }