+++ native/src/main/java/io/ballerina/stdlib/constraint/annotations/TypeAnnotations.java
@@ -28,27 +28,30 @@ import io.ballerina.runtime.api.values.BDecimal;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BString;
 
-import java.util.Set;
+import java.util.List;
+
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_CLOSE_SQUARE_BRACKET;
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_OPEN_SQUARE_BRACKET;
 
 /**
  * Extern functions for validating constraints on types.
  */
 public class TypeAnnotations extends AbstractAnnotations {
 
-    private final Set<String> failedConstraints;
+    private final List<String> failedConstraints;
 
-    public TypeAnnotations(Set<String> failedConstraints) {
+    public TypeAnnotations(List<String> failedConstraints) {
         super(failedConstraints);
         this.failedConstraints = failedConstraints;
     }
 
     @Override
-    public void validate(Object value, AnnotatableType type) {
+    public void validate(Object value, AnnotatableType type, String path) {
         BMap<BString, Object> typeAnnotations = type.getAnnotations();
         Object fieldValue = getFieldValue(value);
-        super.validateAnnotations(typeAnnotations, fieldValue);
+        super.validateAnnotations(typeAnnotations, fieldValue, path);
         if (type instanceof ReferenceType) {
-            validateReferredType(value, (ReferenceType) type);
+            validateReferredType(value, (ReferenceType) type, path);
         }
     }
 
@@ -63,33 +66,35 @@ public class TypeAnnotations extends AbstractAnnotations {
         return value;
     }
 
-    private void validateReferredType(Object value, ReferenceType type) {
+    private void validateReferredType(Object value, ReferenceType type, String path) {
         Type referredType = type.getReferredType();
         if (referredType instanceof AnnotatableType) {
             if (referredType instanceof RecordType) {
                 RecordFieldAnnotations recordFieldAnnotations = new RecordFieldAnnotations(this.failedConstraints);
-                recordFieldAnnotations.validate(value, (AnnotatableType) referredType);
+                recordFieldAnnotations.validate(value, (AnnotatableType) referredType, path);
             } else {
-                validate(value, (AnnotatableType) referredType);
+                validate(value, (AnnotatableType) referredType, path);
             }
         } else if (referredType instanceof ArrayType) {
-            validateArrayType((ArrayType) referredType, (BArray) value);
+            validateArrayType((ArrayType) referredType, (BArray) value, path);
         }
     }
 
     @SuppressWarnings("unchecked")
-    private void validateArrayType(ArrayType referredType, BArray value) {
+    private void validateArrayType(ArrayType referredType, BArray value, String path) {
         Type elementType = referredType.getElementType();
         if (elementType instanceof ReferenceType && elementType instanceof AnnotatableType) {
             if (elementType instanceof RecordType) {
                 RecordFieldAnnotations recordFieldAnnotations = new RecordFieldAnnotations(this.failedConstraints);
                 for (int i = 0; i < value.getLength(); i++) {
                     BMap<BString, Object> map = (BMap<BString, Object>) value.getRefValue(i);
-                    recordFieldAnnotations.validate(map, (AnnotatableType) elementType);
+                    recordFieldAnnotations.validate(map, (AnnotatableType) elementType,
+                            path + SYMBOL_OPEN_SQUARE_BRACKET + i + SYMBOL_CLOSE_SQUARE_BRACKET);
                 }
             } else {
                 for (int i = 0; i < value.getLength(); i++) {
-                    validate(value.getRefValue(i), (AnnotatableType) elementType);
+                    validate(value.getRefValue(i), (AnnotatableType) elementType,
+                            path + SYMBOL_OPEN_SQUARE_BRACKET + i + SYMBOL_CLOSE_SQUARE_BRACKET);
                 }
             }
         }