+++ native/src/main/java/io/ballerina/stdlib/constraint/validators/AbstractLengthValidator.java
@@ -20,33 +20,38 @@ package io.ballerina.stdlib.constraint.validators;
 
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.stdlib.constraint.Constants;
 
+import java.util.List;
 import java.util.Map;
-import java.util.Set;
+
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_LENGTH;
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MAX_LENGTH;
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MIN_LENGTH;
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_SEPARATOR;
 
 /**
  * The abstract class to validate the length related constraints.
  */
 public abstract class AbstractLengthValidator {
 
-    public void validate(BMap<BString, Object> constraints, Object fieldValue, Set<String> failedConstraints) {
+    public void validate(BMap<BString, Object> constraints, Object fieldValue, List<String> failedConstraints,
+                         String path) {
         for (Map.Entry<BString, Object> constraint : constraints.entrySet()) {
             long constraintValue = (long) constraint.getValue();
             switch (constraint.getKey().getValue()) {
-                case Constants.CONSTRAINT_LENGTH:
+                case CONSTRAINT_LENGTH:
                     if (!validateLength(fieldValue, constraintValue)) {
-                        failedConstraints.add(Constants.CONSTRAINT_LENGTH);
+                        failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_LENGTH);
                     }
                     break;
-                case Constants.CONSTRAINT_MIN_LENGTH:
+                case CONSTRAINT_MIN_LENGTH:
                     if (!validateMinLength(fieldValue, constraintValue)) {
-                        failedConstraints.add(Constants.CONSTRAINT_MIN_LENGTH);
+                        failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MIN_LENGTH);
                     }
                     break;
-                case Constants.CONSTRAINT_MAX_LENGTH:
+                case CONSTRAINT_MAX_LENGTH:
                     if (!validateMaxLength(fieldValue, constraintValue)) {
-                        failedConstraints.add(Constants.CONSTRAINT_MAX_LENGTH);
+                        failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MAX_LENGTH);
                     }
                     break;
                 default: