+++ native/src/main/java/io/ballerina/stdlib/constraint/validators/AbstractValueValidator.java
@@ -19,10 +19,15 @@
 package io.ballerina.stdlib.constraint.validators;
 
 import io.ballerina.runtime.api.values.BString;
-import io.ballerina.stdlib.constraint.Constants;
 
+import java.util.List;
 import java.util.Map;
-import java.util.Set;
+
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MAX_VALUE;
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MAX_VALUE_EXCLUSIVE;
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MIN_VALUE;
+import static io.ballerina.stdlib.constraint.Constants.CONSTRAINT_MIN_VALUE_EXCLUSIVE;
+import static io.ballerina.stdlib.constraint.Constants.SYMBOL_SEPARATOR;
 
 /**
  * The abstract class to validate the value related constraints.
@@ -30,26 +35,26 @@ import java.util.Set;
 public abstract class AbstractValueValidator {
 
     public void validate(Map.Entry<BString, Object> constraint, Number fieldValue, Number constraintValue,
-                         Set<String> failedConstraints) {
+                         List<String> failedConstraints, String path) {
         switch (constraint.getKey().getValue()) {
-            case Constants.CONSTRAINT_MIN_VALUE:
+            case CONSTRAINT_MIN_VALUE:
                 if (!validateMinValue(fieldValue, constraintValue)) {
-                    failedConstraints.add(Constants.CONSTRAINT_MIN_VALUE);
+                    failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MIN_VALUE);
                 }
                 break;
-            case Constants.CONSTRAINT_MAX_VALUE:
+            case CONSTRAINT_MAX_VALUE:
                 if (!validateMaxValue(fieldValue, constraintValue)) {
-                    failedConstraints.add(Constants.CONSTRAINT_MAX_VALUE);
+                    failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MAX_VALUE);
                 }
                 break;
-            case Constants.CONSTRAINT_MIN_VALUE_EXCLUSIVE:
+            case CONSTRAINT_MIN_VALUE_EXCLUSIVE:
                 if (!validateMinValueExclusive(fieldValue, constraintValue)) {
-                    failedConstraints.add(Constants.CONSTRAINT_MIN_VALUE_EXCLUSIVE);
+                    failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MIN_VALUE_EXCLUSIVE);
                 }
                 break;
-            case Constants.CONSTRAINT_MAX_VALUE_EXCLUSIVE:
+            case CONSTRAINT_MAX_VALUE_EXCLUSIVE:
                 if (!validateMaxValueExclusive(fieldValue, constraintValue)) {
-                    failedConstraints.add(Constants.CONSTRAINT_MAX_VALUE_EXCLUSIVE);
+                    failedConstraints.add(path + SYMBOL_SEPARATOR + CONSTRAINT_MAX_VALUE_EXCLUSIVE);
                 }
                 break;
             default: